update the dist folder and create api.js file
