+++ src/modules/api.js
@@ -1,0 +1,4 @@
+const baseUrl = 'https://api.tvmaze.com/shows';
+const involvementUrl = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/fZEB2tpIC45iiUMRjYli/';
+
+export { baseUrl, involvementUrl };
