+++ src/style.css
@@ -54,7 +54,6 @@ header #link a {
 }
 
 #home .container {
-  /* border: 1px solid red; */
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -127,7 +126,7 @@ h2 {
 
 .comment {
   width: 100%;
-  height: 100%;
+  height: auto;
   position: fixed;
   top: 0;
   background: rgba(151, 148, 147, 70%);
@@ -136,7 +135,7 @@ h2 {
   flex-direction: column;
   align-items: center;
   gap: 15px;
-  padding: 30px;
+  padding: 0px 30px;
 }
 
 .properties {
@@ -165,7 +164,7 @@ input {
 textarea {
   padding: 5px;
   width: 300px;
-  height: 100px;
+  height: 60px;
 }
 
 .hide {
